+++ open-starter-crawler/src/main/java/com/saucesubfresh/starter/crawler/handler/DefaultResultHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.saucesubfresh.starter.crawler.executor.result;
+package com.saucesubfresh.starter.crawler.handler;
 
 
 import com.saucesubfresh.starter.crawler.domain.FieldExtractor;
