--- open-starter-crawler/src/main/java/com/saucesubfresh/starter/crawler/plugin/InterceptorChain.java
@@ -1,59 +1,0 @@
-/*
- * Copyright © 2022 organization SauceSubFresh
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.saucesubfresh.starter.crawler.plugin;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * @author lijunping
- */
-public class InterceptorChain {
-    /**
-     * 拦截器集合
-     */
-    private final List<Interceptor> interceptors = new ArrayList<>();
-
-    /**
-     * 装载拦截器
-     */
-    public Object pluginAll(Object target) {
-        for (Interceptor interceptor : interceptors) {
-            target = interceptor.plugin(target);
-        }
-        return target;
-    }
-
-    /**
-     * 添加拦截器
-     *
-     * @param interceptor 拦截器
-     */
-    public void addInterceptor(Interceptor interceptor) {
-        interceptors.add(interceptor);
-    }
-
-
-    /**
-     * 获取所有拦截器
-     *
-     * @return 拦截器集合
-     */
-    public List<Interceptor> getInterceptors() {
-        return Collections.unmodifiableList(interceptors);
-    }
-}
