--- open-starter-crawler/src/main/java/com/saucesubfresh/starter/crawler/plugin/Invocation.java
@@ -1,51 +1,0 @@
-/*
- * Copyright © 2022 organization SauceSubFresh
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.saucesubfresh.starter.crawler.plugin;
-
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-
-/**
- * @author lijunping
- */
-public class Invocation {
-
-    private final Object target;
-    private final Method method;
-    private final Object[] args;
-
-    public Invocation(Object target, Method method, Object[] args) {
-        this.target = target;
-        this.method = method;
-        this.args = args;
-    }
-
-    public Object getTarget() {
-        return target;
-    }
-
-    public Method getMethod() {
-        return method;
-    }
-
-    public Object[] getArgs() {
-        return args;
-    }
-
-    public Object proceed() throws InvocationTargetException, IllegalAccessException {
-        return method.invoke(target, args);
-    }
-}
