--- open-starter-crawler/src/main/java/com/saucesubfresh/starter/crawler/plugin/Plugin.java
@@ -1,87 +1,0 @@
-/*
- * Copyright © 2022 organization SauceSubFresh
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.saucesubfresh.starter.crawler.plugin;
-
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.Method;
-import java.lang.reflect.Proxy;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-
-/**
- * @author lijunping
- */
-public class Plugin implements InvocationHandler {
-
-    private final Object target;
-    private final Interceptor interceptor;
-    private final Map<Class<?>, Set<Method>> signatureMap;
-
-    private Plugin(Object target, Interceptor interceptor, Map<Class<?>, Set<Method>> signatureMap) {
-        this.target = target;
-        this.interceptor = interceptor;
-        this.signatureMap = signatureMap;
-    }
-
-    /**
-     * 包装代理
-     *
-     * @param interceptor 拦截器
-     */
-    public static Object wrap(Object t, Interceptor interceptor) {
-        Map<Class<?>, Set<Method>> signatureMap = getSignatureMap(interceptor);
-        return Proxy.newProxyInstance(
-                t.getClass().getClassLoader(),
-                t.getClass().getInterfaces(),
-                new Plugin(t, interceptor, signatureMap));
-    }
-
-    @Override
-    public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-        try {
-            Set<Method> methods = signatureMap.get(method.getDeclaringClass());
-            if (methods != null && methods.contains(method)) {
-                return interceptor.intercept(new Invocation(target, method, args));
-            }
-            return method.invoke(target, args);
-        } catch (Exception e) {
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-    private static Map<Class<?>, Set<Method>> getSignatureMap(Interceptor interceptor) {
-        Intercepts interceptsAnnotation = interceptor.getClass().getAnnotation(Intercepts.class);
-        // 检查类是否被@Intercepts标记
-        if (interceptsAnnotation == null) {
-            throw new RuntimeException("No @Intercepts annotation was found in interceptor " + interceptor.getClass().getName());
-        }
-        Signature[] sings = interceptsAnnotation.value();
-        Map<Class<?>, Set<Method>> signatureMap = new HashMap<>();
-        // 检查被@Signature标记的方法是否存在
-        for (Signature sig : sings) {
-            Set<Method> methods = signatureMap.computeIfAbsent(sig.type(), k -> new HashSet<>());
-            try {
-                Method method = sig.type().getMethod(sig.method(), sig.args());
-                methods.add(method);
-            } catch (NoSuchMethodException e) {
-                throw new RuntimeException("Could not find method on " + sig.type() + " named " + sig.method() + ". Cause: " + e, e);
-            }
-        }
-        return signatureMap;
-    }
-}
